+++ Develappers.BillomatNet/CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - implemented service for CRD operations on one client tag
 - implemented service for CR operations on one invoice comment
 - implemented service for R operations on invoice comments
+- implemented service for CR operations on one invoice payment
+- implemented service for R operations on invoice payments
 
 ### Changed
 - GetList can throw an exception when not authorized now